New Design flaws solution

+++ src/main/java/com/creativityskills/jotech/model/CashDrawer.java
@@ -1,15 +1,14 @@
 package com.creativityskills.jotech.model;
 
 public class CashDrawer {
-    private long id;
+    private int id;
     private Denomination denomination;
     private long count;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
-
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
